refactor: Move all source code into 'src/'
Now the index.ts doesn't live among the configuration files, and is
instead treated as a source file (which it has always been, of course).
