+++ src/config.ts
@@ -1,11 +1,11 @@
-import { GraaOctokit, RepoOfAuthenticatedUser } from './types.js'
-import { tryReadFileAsUtf8 } from './content.js'
+import { GraaOctokit, RepoOfAuthenticatedUser } from './github/types.js'
+import { tryReadFileAsUtf8 } from './github/content.js'
 import { Infer, object, optional, record, string, validate } from 'superstruct'
 import * as YAML from 'yaml'
 import { RepoConfigError } from './errors.js'
-import { baseConfig } from '../configs/base.js'
-import { jsLibConfig } from '../configs/js-lib.js'
-import { jsAppConfig } from '../configs/js-app.js'
+import { baseConfig } from './configs/base.js'
+import { jsLibConfig } from './configs/js-lib.js'
+import { jsAppConfig } from './configs/js-app.js'
 import { getAutomation } from './automation.js'
 import { assertAutomationOptions } from './validation.js'
