+++ src/configs/js-lib.ts
@@ -1,8 +1,8 @@
-import { PartialConfig } from '../lib/config.js'
+import { PartialConfig } from '../config.js'
 
 const renovateJson = `{
   "extends": [
-    "config:base"
+    "config:js-lib"
   ],
   "labels": ["dependencies"],
   "packageRules": [
@@ -14,12 +14,9 @@ const renovateJson = `{
 }
 `
 
-export const baseConfig: PartialConfig = {
+export const jsLibConfig: PartialConfig = {
+  extends: 'config:base',
   automations: {
-    reconfigure: {
-      'delete-branch-on-merge': true
-    },
-    'license-date': {},
     files: {
       'renovate.json': renovateJson
     }
