+++ src/log.ts
@@ -1,4 +1,4 @@
-import { RepoOfAuthenticatedUser } from './types.js'
+import { RepoOfAuthenticatedUser } from './github/types.js'
 
 export interface Log {
   info: (message: any, ...args: any[]) => void
