+++ tsconfig.json
@@ -6,9 +6,7 @@
     "strict": true,
     "outDir": "./dist"
   },
-  "exclude": [
-    "test",
-    "dist",
-    "node_modules"
+  "include": [
+    "src"
   ]
 }
